Move compiletest test steps and related helpers into `compiletest_suites` module
